clean up and comments
