+++ app/src/app/components/sidebar/PopulationPanel/PopulationPanel.tsx
@@ -8,7 +8,8 @@ import {useMapStore} from '@store/mapStore';
 import {PopulationChart} from './PopulationChart/PopulationChart';
 import {PopulationPanelOptions} from './PopulationPanelOptions';
 import {LockClosedIcon, LockOpen2Icon} from '@radix-ui/react-icons';
-import { useDemography, useSummaryStats } from '@/app/utils/demography/demographyCache';
+import { useDemography } from '@/app/hooks/useDemography';
+import { useSummaryStats } from '@/app/hooks/useSummaryStats';
 
 const maxNumberOrderedBars = 40; // max number of zones to consider while keeping blank spaces for missing zones
